Calculator 3

+++ Calculator.js
@@ -1,19 +1,16 @@
 function calc() {
-    var a = parseInt(document.querySelector("#value1").value);
-    var b = parseInt(document.querySelector("#value2").value);
+    var a = parseFloat(document.querySelector("#value1").value);
+    var b = parseFloat(document.querySelector("#value2").value);
     var op = document.querySelector("#operator").value;
     var calculate;
 
-    if (op == "add"){
-        calculate = a +b;
-    }
-    else if (op == "min"){
+    if (op == "add") {
+        calculate = a + b;
+    } else if (op == "min") {
         calculate = a - b;
-    }
-    else if (op == "div"){
+    } else if (op == "div") {
         calculate = a / b;
-    }
-    else if (op == "mul"){
+    } else if (op == "mul") {
         calculate = a * b;
     }
